+++ src/components/BankTransactionMobileList/constants.ts
@@ -1,0 +1,1 @@
+export const PersonalCategories = ['PERSONAL_INCOME', 'PERSONAL_EXPENSES']
