+++ src/styles/bank_transactions_mobile_list.scss
@@ -132,7 +132,6 @@
     }
   }
 
-  // @TODO temp
   .Layer__select__control {
     width: 150px;
     font-size: 12px;
